Remove interface class for subscription service
